Model Loader
Added a loader which will fill the vertex and index buffers of an object with data loaded from an OBJ file.

The cube we have had hardcoded into the code is also now an object called startingCube.obj.
Adjusted the scale of the box object so it is nicer to look at.