Added Texture Loading
The load model function now loads a diffuse texture for the model if it can be found. 
The different parts of the model are now loaded into unique sub-meshes (children of the current Object). Submesh objects do nothing when initalized and update their model matrix to match their parent every frame.