Added Command Line Arguments to Config File
The program now checks the config file for command line arguments (vertex, fragment, and model paths) if they aren't set on the command line. 
The json config file is now stored in the arguments class and can be accessed with the getConfig() function.
The config file has a default location (config.json) set.