Improved CircularBuffer, Fixed Diagonal Bug
The circular buffer now no longer requires an extra blank slot (at the cost of iterators having an extra bool (which is less overhead for any type of complex type)).

The add chunk functions where adding to different ends of the buffer for different directions which was causing a big hole to appear in the world after repeated diagonal steps.