Chunk Finalization and Pipeline Withdrawl
When a chunk is removed from the world it is marked as freed. If any of the chunk generation steps then try to access the freed chunk, they notice and remove it from the generation pipeline.

secure_circular_buffer has been refactored into finalizeable_circular_buffer to support this change. Whenever a chunk would be swapped out, it is marked as freed.